Update Get-ScriptBlockText to handle missing #text property
In this case, it just throws an error at you when it finds the `ScriptBlockText` node in the event data, but if the `#text` property isn't there it errors out.

This fixes that - didn't really impact speed that much either. Just added checks if the node exists, then if the text is null or not.